+++ pytket/docs/changelog.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Minor new features:
+
+* Add optional parameter to QASM conversion methods to set the maximum allowed
+  width of classical registers (default 32).
+* New ``OpType.CS`` and ``OpType.CSdg``.
+
 1.21.0 (October 2023)
 ---------------------